+++ MacroC-ClientPart/MacroC-ClientPart/View/AuthView/LoginView.swift
@@ -1,88 +1,10 @@
-////
-////  LoginView.swift
-////  MacroC-ClientPart
-////
-////  Created by Kimdohyun on 2023/10/05.
-////
 //
-//import SwiftUI
+//  LoginView.swift
+//  MacroC-ClientPart
 //
-//struct LoginView: View {
+//  Created by Kimdohyun on 2023/10/05.
 //
-//    //MARK: -1.PROPERTY
-//    @StateObject var viewModel = LoginViewModel()
-//
-//    //MARK: -2.BODY
-//    var body: some View {
-//
-//            VStack(alignment: .center) {
-//                Spacer()
-//                Spacer()
-//                textFieldList
-//                signInbutton
-//                AppleSigninButton()
-//                Spacer()
-//            }
-//            .padding()
-//            .background(backgroundView().hideKeyboardWhenTappedAround())
-//    }
-//}
-//
-//
-////MARK: -3.PREVIEW
-//#Preview {
-//    LoginView()
-//}
-//
-////MARK: -4.EXTENSION
-//extension LoginView {
-//    var textFieldList: some View {
-//        VStack(spacing: 5){
-//            TextField("Email", text: $viewModel.email)
-//                .padding(13)
-//                .background(.ultraThinMaterial)
-//                .cornerRadius(6)
-//
-//            TextField("Password", text: $viewModel.password)
-//                .padding(13)
-//                .background(.ultraThinMaterial)
-//                .cornerRadius(6)
-//        }
-//    }
-//
-////    var signInbutton: some View {
-////        Button {
-//////            viewModel.()
-////        } label: {
-////            HStack{
-////                Spacer()
-////                Text("Sign In").fontWeight(.semibold)
-////                Spacer()
-////            }
-////            .padding()
-////            .background(.ultraThinMaterial)
-////            .cornerRadius(6)
-////
-////        }.padding(.top, 30)
-////    }
-//    var signInbutton: some View {
-//        Button {
-////            viewModel.signIn()
-//        } label: {
-//            HStack{
-//                Spacer()
-//                Text("Sign In").fontWeight(.semibold)
-//                Spacer()
-//            }
-//            .padding()
-//            .background(.ultraThinMaterial)
-//            .background(viewModel.email.isEmpty || viewModel.password.isEmpty ?  Color.black : Color.blue.opacity(0.7))
-//            .cornerRadius(6)
-//        }.padding(.top, 30)
-//            .disabled(viewModel.email.isEmpty || viewModel.password.isEmpty)
-//
-//    }
-//}
+
 import SwiftUI
 import Firebase
 import FirebaseAuth
@@ -91,16 +13,13 @@ import AuthenticationServices
 
 struct LoginView: View {
 
-    
     @EnvironmentObject var awsService : AwsService
+    @StateObject var loginData = LoginViewModel()
+    
     @State private var isLoggedin: Bool = false
     @State private var userUID: String = ""
     @State private var showLoginUI: Bool = false
     
-    @StateObject var loginData = LoginViewModel()
-    
-    
-
     var body: some View {
         VStack {
             Text(isLoggedin ? "Logged in: \(userUID)" : "Not Logged in")
